+++ structures/Aggregatestats.h
@@ -1,0 +1,21 @@
+//
+// Created by shira on 5/20/2024.
+//
+#include <string>
+#include <vector>
+#include "../imports/json.hpp"
+#include <fstream>
+#include <iostream>
+
+#ifndef TETR_IO_VISULIZER_AGGREGATESTATS_H
+#define TETR_IO_VISULIZER_AGGREGATESTATS_H
+
+class Aggregatestats{
+public:
+    double apm;
+    double pps;
+    double vsscore;
+};
+
+
+#endif //TETR_IO_VISULIZER_AGGREGATESTATS_H
