+++ structures/AvgTracking.h
@@ -1,0 +1,18 @@
+//
+// Created by shira on 5/20/2024.
+//
+#include <string>
+#include <vector>
+#include "../imports/json.hpp"
+#include <fstream>
+#include <iostream>
+#ifndef TETR_IO_VISULIZER_AVGTRACKING_H
+#define TETR_IO_VISULIZER_AVGTRACKING_H
+
+class AvgTracking {
+public:
+    std::vector<double> arr;
+};
+
+
+#endif //TETR_IO_VISULIZER_AVGTRACKING_H
