+++ structures/BoardV.h
@@ -1,0 +1,19 @@
+//
+// Created by shira on 5/20/2024.
+//
+#include <string>
+#include <vector>
+#include "../imports/json.hpp"
+#include <fstream>
+#include <iostream>
+#ifndef TETR_IO_VISULIZER_BOARDV_H
+#define TETR_IO_VISULIZER_BOARDV_H
+
+
+class BoardV{
+public:
+    std::vector<std::string> row;
+};
+
+
+#endif //TETR_IO_VISULIZER_BOARDV_H
