+++ structures/Clears.h
@@ -1,0 +1,34 @@
+//
+// Created by shira on 5/20/2024.
+//
+#include <string>
+#include <vector>
+#include "../imports/json.hpp"
+#include <fstream>
+#include <iostream>
+#ifndef TETR_IO_VISULIZER_CLEARS_H
+#define TETR_IO_VISULIZER_CLEARS_H
+
+
+class Clears {
+public:
+    int singles;
+    int doubles;
+    int triples;
+    int quads;
+    int pentas;
+    int realtspins;
+    int minitspins;
+    int minitspinsingles;
+    int tspinsingles;
+    int minitspindoubles;
+    int tspindoubles;
+    int tspintriples;
+    int tspinquads;
+    int tspinpentas;
+    int allclear;
+};
+
+
+
+#endif //TETR_IO_VISULIZER_CLEARS_H
