+++ structures/ConstantsOverrides.h
@@ -1,0 +1,20 @@
+//
+// Created by shira on 5/20/2024.
+//
+#include <string>
+#include <vector>
+#include "../imports/json.hpp"
+#include <fstream>
+#include <iostream>
+
+#ifndef TETR_IO_VISULIZER_CONSTANTS_OVERRIDES_H
+#define TETR_IO_VISULIZER_CONSTANTS_OVERRIDES_H
+
+
+class Constants_overrides {
+public:
+
+};
+
+
+#endif //TETR_IO_VISULIZER_CONSTANTS_OVERRIDES_H
