+++ structures/Enemy.h
@@ -1,0 +1,17 @@
+//
+// Created by shira on 5/20/2024.
+//
+#include <string>
+#include <vector>
+#include "../imports/json.hpp"
+#include <fstream>
+#include <iostream>
+#ifndef TETR_IO_VISULIZER_ENEMY_H
+#define TETR_IO_VISULIZER_ENEMY_H
+
+class Enemy{
+public:
+
+};
+
+#endif //TETR_IO_VISULIZER_ENEMY_H
