+++ structures/Finesse.h
@@ -1,0 +1,20 @@
+//
+// Created by shira on 5/20/2024.
+//
+#include <string>
+#include <vector>
+#include "../imports/json.hpp"
+#include <fstream>
+#include <iostream>
+#ifndef TETR_IO_VISULIZER_FINESSE_H
+#define TETR_IO_VISULIZER_FINESSE_H
+
+class Finesse{
+public:
+    int combo;
+    int faults;
+    int perfectpieces;
+};
+
+
+#endif //TETR_IO_VISULIZER_FINESSE_H
