+++ structures/Garbage.h
@@ -1,0 +1,21 @@
+//
+// Created by shira on 5/20/2024.
+//
+#include <string>
+#include <vector>
+#include "../imports/json.hpp"
+#include <fstream>
+#include <iostream>
+#ifndef TETR_IO_VISULIZER_GARBAGE_H
+#define TETR_IO_VISULIZER_GARBAGE_H
+
+class Garbage{
+public:
+    int sent;
+    int received;
+    int attack;
+    int cleared;
+};
+
+
+#endif //TETR_IO_VISULIZER_GARBAGE_H
