+++ structures/Handling.h
@@ -1,0 +1,25 @@
+//
+// Created by shira on 5/20/2024.
+//
+#include <string>
+#include <vector>
+#include "../imports/json.hpp"
+#include <fstream>
+#include <iostream>
+#ifndef TETR_IO_VISULIZER_HANDLING_H
+#define TETR_IO_VISULIZER_HANDLING_H
+
+
+class Handling {
+public:
+    int arr;
+    int das;
+    int dcd;
+    int sdf;
+    bool safelock;
+    bool cancel;
+    bool may20g;
+};
+
+
+#endif //TETR_IO_VISULIZER_HANDLING_H
