+++ structures/Hold.h
@@ -1,0 +1,21 @@
+//
+// Created by shira on 5/20/2024.
+//
+#include <string>
+#include <vector>
+#include "../imports/json.hpp"
+#include <fstream>
+#include <iostream>
+#ifndef TETR_IO_VISULIZER_HOLD_H
+#define TETR_IO_VISULIZER_HOLD_H
+
+
+class Hold{
+public:
+    std::string piece;
+    bool locked;
+};
+
+
+
+#endif //TETR_IO_VISULIZER_HOLD_H
