+++ structures/Killer.h
@@ -1,0 +1,20 @@
+//
+// Created by shira on 5/20/2024.
+//
+#include <string>
+#include <vector>
+#include "../imports/json.hpp"
+#include <fstream>
+#include <iostream>
+#ifndef TETR_IO_VISULIZER_KILLER_H
+#define TETR_IO_VISULIZER_KILLER_H
+
+
+class Killer{
+public:
+    std::string gameid;
+    std::string name;
+    std::string type;
+};
+
+#endif //TETR_IO_VISULIZER_KILLER_H
