+++ structures/Minoskin.h
@@ -1,0 +1,27 @@
+//
+// Created by shira on 5/20/2024.
+//
+#include <string>
+#include <vector>
+#include "../imports/json.hpp"
+#include <fstream>
+#include <iostream>
+#ifndef TETR_IO_VISULIZER_MINOSKIN_H
+#define TETR_IO_VISULIZER_MINOSKIN_H
+
+
+class Minoskin {
+public:
+    std::string z;
+    std::string l;
+    std::string o;
+    std::string s;
+    std::string i;
+    std::string j;
+    std::string t;
+    std::string other;
+};
+
+
+
+#endif //TETR_IO_VISULIZER_MINOSKIN_H
